LayoutTests/media/media-source/media-source-multiple-trackBuffers-buffered-in-ended-state.html

<!DOCTYPE html>
<html>
<head>
    <title>mock-media-source</title>
    <script src="mock-media-source.js"></script>
    <script src="../video-test.js"></script>
    <script>
    var source;
    var sourceBuffer;
    var init;
    var samples;

    if (window.internals)
        internals.initializeMockMediaSource();

    function runTest() {
        findMediaElement();
        source = new MediaSource();
        source.addEventListener('sourceopen', startLoad);
        run('video.src = URL.createObjectURL(source)');
    }

    function startLoad() {
        sourceBuffer = source.addSourceBuffer('video/mock; codecs="mock"');
        waitForEventOn(sourceBuffer, 'updateend', loadSamples1, false, true);

        // Make an init segment with 1 video track and 1 audio track
        init = makeAInit(0, [
            makeATrack(1, 'mock', TRACK_KIND.VIDEO),
            makeATrack(2, 'mock', TRACK_KIND.AUDIO),
            ]);
        testExpected('sourceBuffer.updating', false);
        run('sourceBuffer.appendBuffer(init);');
    }

    function loadSamples1() {
        samples = concatenateSamples([
            makeASample(3, 3, 1, 1, SAMPLE_FLAG.SYNC),
            makeASample(4, 4, 1, 1, SAMPLE_FLAG.NONE),
            makeASample(5, 5, 1, 1, SAMPLE_FLAG.NONE),
        ]);
        waitForEventOn(sourceBuffer, 'updateend', loadSamples2, false, true);
        run('sourceBuffer.appendBuffer(samples)');
    }

    function loadSamples2() {
        testExpected('sourceBuffer.buffered.length', 0);

        samples = concatenateSamples([
            makeASample(0, 0, 1, 2, SAMPLE_FLAG.SYNC),
            makeASample(1, 1, 1, 2, SAMPLE_FLAG.SYNC),
            makeASample(2, 2, 1, 2, SAMPLE_FLAG.SYNC),
            makeASample(3, 3, 1, 2, SAMPLE_FLAG.SYNC),

            makeASample(7, 7, 1, 2, SAMPLE_FLAG.SYNC),
            makeASample(8, 8, 1, 2, SAMPLE_FLAG.SYNC),
            makeASample(9, 9, 1, 2, SAMPLE_FLAG.SYNC),
        ]);
        waitForEventOn(sourceBuffer, 'updateend', loadSamples3, false, true);
        run('sourceBuffer.appendBuffer(samples)');
    }

    function loadSamples3() {
        testExpected('sourceBuffer.buffered.length', 1);
        testExpected('sourceBuffer.buffered.start(0).toFixed(3)', 3);
        testExpected('sourceBuffer.buffered.end(0).toFixed(3)', 4);

        waitForEventOn(source, 'sourceended', ended, false, true);
        waitForEventOn(sourceBuffer, 'updateend', finishTest, false, true);
        run('source.endOfStream()');
    }

    function ended() {
        testExpected('source.readyState', "ended");
        testExpected('sourceBuffer.buffered.length', 2);
        testExpected('sourceBuffer.buffered.start(1).toFixed(3)', 7);
        testExpected('sourceBuffer.buffered.end(1).toFixed(3)', 10);
    }

    function finishTest() {
        testExpected('source.readyState', "ended");
        run('sourceBuffer.timestampOffset = 0;'); // this opens MediaSource
        testExpected('source.readyState', "open");
        testExpected('sourceBuffer.buffered.length', 1);
        testExpected('sourceBuffer.buffered.start(0).toFixed(3)', 3);
        testExpected('sourceBuffer.buffered.end(0).toFixed(3)', 4);
        endTest();
    }

    </script>
</head>
<body onload="runTest()">
    <video></video>
</body>
</html>
